+++ packages/substrate-bindings/src/codecs/scale/Binary.ts
@@ -1,37 +1,68 @@
-import { Bytes, Codec, Decoder, Encoder, createCodec } from "scale-ts"
-import { fromHex, toHex } from "@polkadot-api/utils"
-import { HexString } from "./Hex"
+import {
+  Bytes,
+  Codec,
+  Decoder,
+  Encoder,
+  Tuple,
+  compact,
+  createCodec,
+} from "scale-ts"
+import { fromHex, mergeUint8, toHex } from "@polkadot-api/utils"
+import type { HexString } from "./Hex"
 
 const textEncoder = new TextEncoder()
 const textDecoder = new TextDecoder()
+const opaqueBytesDec = Tuple(compact, Bytes(Infinity))[1]
 
 export class Binary {
   #bytes: Uint8Array
+  #opaqueBytes: Uint8Array | null = null
   #hex: HexString | null = null
+  #opaqueHex: HexString | null = null
   #str: string | null = null
 
-  constructor(data: Uint8Array) {
-    this.#bytes = data
+  constructor(data: Uint8Array, opaque = false) {
+    if (opaque) {
+      try {
+        const [len, bytes] = opaqueBytesDec(data)
+        if (len === bytes.length) {
+          this.#bytes = bytes
+          this.#opaqueBytes = data
+          return
+        }
+      } catch (_) {}
+      throw new Error("Invalid opaque bytes")
+    } else this.#bytes = data
   }
 
-  asText = () =>
-    this.#str === null
-      ? (this.#str = textDecoder.decode(this.#bytes))
-      : this.#str
+  asText = () => (this.#str ??= textDecoder.decode(this.#bytes))
 
-  asHex = () =>
-    this.#hex === null ? (this.#hex = toHex(this.#bytes)) : this.#hex
+  asHex = () => (this.#hex ||= toHex(this.#bytes))
+  asOpaqueHex = () => (this.#opaqueHex ||= toHex(this.asBytes()))
 
   asBytes = () => this.#bytes
+  asOpaqueBytes = () =>
+    (this.#opaqueBytes ||= mergeUint8(
+      compact[0](this.#bytes.length),
+      this.#bytes,
+    ))
 
   static fromText(input: string): Binary {
     return new this(textEncoder.encode(input))
   }
+
   static fromHex(input: HexString): Binary {
     return new this(fromHex(input))
   }
+  static fromOpaqueHex(input: HexString): Binary {
+    return new this(fromHex(input), true)
+  }
+
   static fromBytes(input: Uint8Array): Binary {
     return new this(input)
+  }
+  static fromOpaqueBytes(input: Uint8Array): Binary {
+    return new this(input, true)
   }
 }