+++ packages/client/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - `pjs-signer`: Remove option byte from `CheckMetadataHash`
 
+### Added
+
+- New `Binary` methods for supporting "`opaque`" binary data. [#675](https://github.com/polkadot-api/polkadot-api/pull/675).
+
 ## 1.1.0 - 2024-08-28
 
 ### Added